2.0.0 > 2.5.0 Rewrite update!
Finally added animations on the expandable lists!
Fixed a good amount of info issues thanks to a contributor
Added waveland as it\'s own technique
Made the entire app modular, if anyone wants to contribute it is much easier :)
Fixed the scroll bar on terms
Improved icon a bit
I will now be focusing on the iOS port!